Fix lazy-init of LibLazyCrafting
It was re-init-ing over and over again, which was good in that it actually returned a useful library, but bad in that the re-init was a waste of time.

Note that the data member WriteWorthy.LibLazyCrafting MUST be initialized to nil, not some LibStub return, since a non-nil initial value prevents proper lazy initialization from noticing a lack of proper initialization. Oops.

Crank up the log size, too. With enough rows, my initialization events are falling off the end of the log.