LibLazyCrafting 1.3 Official
Fixed various typos in the Alchemy file.
Went back to GetTimeStamp() in the smithing file (as it should have been. Sorry!)
Added back timesToMake as an optional parameter in the Alchemy file.
Replaced the body of the crafting complete function with the stackableCraftingComplete function in the alchemy file.
Removed the FindIncreasedSlotIndex from the Alchemy file, as it was already contained in the main LLC file.