LLC isItemCraftable hooks (partial)
LibLazyCrafting lacks working preflight checks for Alchemy and Provisioning, and will erroneously attempt to craft requests that lack required materials. For Alchemy, this is annoying but survivable. For Provisioning this causes an infinite loop and disconnect from server.

Attempt to fix it from outside of LibLazyCrafting, patching in our own preflight hooks. Avoid requiring a change to LibLazyCrafting if at all possible.