WritWorthy.Smithing.Discover()
To get lists of motif/mat/item/trait/whatever values for each of the 6 writ slots in a master writ's item_link.

Might as well make this loop a permanent part of WritWorthy so that I can re-run this next year without as much thinking and hand-typing for loops into the chat window.