LibLazyCrafting: update to 2.1
Dolgubon's big release for Summerset.

Version 2.1: 

Improvement level checking now uses textures, which is non language dependent.
Now checks the crafting skill for Blacksmithing, Clothing, and Woodworking to see if the user has a high enough skill to make an item.
Compiling mats now works for Jewelry Crafting.
Jewelry Crafting will no longer fail due to lack of style mats.
Added a new field to the Craft Request tables, craftNow. By default not set (i.e. set to nil) if it is true, the library will craft the item when at a station. It can only be set to true while at a station, and if the user exits the station it will be set back to false automatically. CraftAllItems and CraftItem now change this new field, and their old functionality has been taken over by the new functions SetAllAutoCraft and SetAutoCraft. 
Added craftInteract to the function table. This is the function which is called when the player interacts with a crafting station. An addon can call it to force the library to rescan the queue and see if there is something it can craft.
Enchanting file added back in.