Start dropping Dolgubon-proprietary setId
Check LLC's internal setId table: if it looks like the old, 1-based table, use proprietary setId values. If not, then use the ESO client values.

Dolgubon pushed a prerelease version of LLC to GitHub that uses the ESO values. This change allows WritWorthy to work with that prerelease version. While also not breaking WritWorthy when working with the older LLC version.