Retain queued status display even when filtered out
Because typing in the filter field does not dequeue any writs that fail the filter criteria. Nor would I want it to, since the first keystroke would start a filter of < 10 gpv, which disables all writs and would be reallly annoying if it dequeued everything I just spent 3 minutes carefully hand-checking and queuing.