Rename WWMUIL → WWMUILContainer
This control is NOT a list. It is the outer shell that _contains_ the header and list.

Also stop calling the wide control atop the list the "header". "header" too often refers to the inner cells atop each column, with labels and sort indicators. Call it "header _container_" and then the little elements inside are header _cell_ controls.